+++ examples/test_fixtures/alb/main.tf
@@ -4,7 +4,7 @@ terraform {
 
 provider "aws" {
   region  = "${var.region}"
-  version = "~> 1.0.0"
+  version = "~> 1.1.0"
 }
 
 provider "template" {